authenticate modify, delete

+++ src/main/java/com/choong/spr/controller/BoardController.java
@@ -65,32 +65,44 @@ public class BoardController {
 	}
 	
 	@PostMapping("modify")
-	public String modify(BoardDto dto, RedirectAttributes rttr) {
-		boolean success = service.updateBoard(dto);
-		
-		if (success) {
-			rttr.addFlashAttribute("message", "글이 수정되었습니다.");
+	public String modify(BoardDto dto, Principal principal, RedirectAttributes rttr) {
+		BoardDto oldBoard = service.getBoardById(dto.getId());
+		if (oldBoard.getMemberId().equals(principal.getName())) {
+			boolean success = service.updateBoard(dto);
+			
+			if (success) {
+				rttr.addFlashAttribute("message", "글이 수정되었습니다.");
+			} else {
+				rttr.addFlashAttribute("message", "글이 수정되지 않았습니다.");
+			}
 		} else {
-			rttr.addFlashAttribute("message", "글이 수정되지 않았습니다.");
+			rttr.addFlashAttribute("message", "권한이 없습니다.");
 		}
-		
 		rttr.addAttribute("id", dto.getId());
 		
 		return "redirect:/board/get";
 	}
 	
 	@PostMapping("remove")
-	public String remove(BoardDto dto, RedirectAttributes rttr) {
+	public String remove(BoardDto dto, Principal principal, RedirectAttributes rttr) {
 		
-		boolean success = service.deleteBoard(dto.getId());
+		// 게시물 정보 얻고
+		BoardDto oldBoard = service.getBoardById(dto.getId());
+		// 게시물 작성자(memberId)와 principal의 name과 비교해서 같을 때만 진행.
+		if (oldBoard.getMemberId().equals(principal.getName())) {
+			boolean success = service.deleteBoard(dto.getId());
 		
-		if (success) {
-			rttr.addFlashAttribute("message", "글이 삭제 되었습니다.");
+			if (success) {
+				rttr.addFlashAttribute("message", "글이 삭제 되었습니다.");
+				
+			} else {
+				rttr.addFlashAttribute("message", "글이 삭제 되지않았습니다.");
+			}
 			
-		} else {
-			rttr.addFlashAttribute("message", "글이 삭제 되지않았습니다.");
+		} else { 
+			rttr.addFlashAttribute("message", "권한이 없습니다.");
 		}
-		
+		// 아니면 리턴.
 		return "redirect:/board/list";
 	}
 }
